Update jampy.py
Fixed the bugs from earlier. Revamped the "next_char()" function and added a "debug" mode that interacts with the shell so if you want to see the token output going forward. I also just barely started the Parser class, although it obviously does nothing right now. Finally done with the tokenizer, for now. I would still like to redo some of the methods to be more efficient in the future. I will also probably change a few other things, such as how parenthesis and things work.